Dmg roll, difficulty variable, transparency, Activate/Destroy method
Damage roll:
The damage roll is the one used throughout the project
Difficulty:
An adjustable difficulty variable
Transparency:
This method calculates the distance between the player and the object,
if it meets the requirement, the object is then rendered for the player
to see
Activate/Destroy:
this method is activated through collision. When the player enters the
trap the player gameobject is constrained on the X and Z axis, but can
rotate for 3 seconds. The damage roll method is applied, and after 3
seconds the trap releases the player gameobject and then destroys
itself.